+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/resolver/DataSourceConstraint.java
@@ -54,6 +54,54 @@ public class DataSourceConstraint {
         ).collect(Collectors.toSet()));
     }
 
+    /**
+     * Constructor.
+     *
+     * @param requestDimensions  Dimensions contained in request
+     * @param filterDimensions  Filtered dimensions
+     * @param metricDimensions  Metric related dimensions
+     * @param metricNames  Names of metrics
+     * @param allDimensions  Set of all dimension objects
+     * @param allDimensionNames  Set of all dimension names
+     * @param allColumnNames  Set of all column names
+     * @param apiFilters  Map of dimension to its set of API filters
+     */
+    protected DataSourceConstraint(
+            Set<Dimension> requestDimensions,
+            Set<Dimension> filterDimensions,
+            Set<Dimension> metricDimensions,
+            Set<String> metricNames,
+            Set<Dimension> allDimensions,
+            Set<String> allDimensionNames,
+            Set<String> allColumnNames,
+            Map<Dimension, Set<ApiFilter>> apiFilters
+    ) {
+        this.requestDimensions = requestDimensions;
+        this.filterDimensions = filterDimensions;
+        this.metricDimensions = metricDimensions;
+        this.metricNames = metricNames;
+        this.allDimensions = allDimensions;
+        this.allDimensionNames = allDimensionNames;
+        this.allColumnNames = allColumnNames;
+        this.apiFilters = apiFilters;
+    }
+
+    /**
+     * Copy Constructor.
+     *
+     * @param dataSourceConstraint  The data source constraint to copy from
+     */
+    protected DataSourceConstraint(DataSourceConstraint dataSourceConstraint) {
+        this.requestDimensions = dataSourceConstraint.getRequestDimensions();
+        this.filterDimensions = dataSourceConstraint.getFilterDimensions();
+        this.metricDimensions = dataSourceConstraint.getMetricDimensions();
+        this.metricNames = dataSourceConstraint.getMetricNames();
+        this.apiFilters = dataSourceConstraint.getApiFilters();
+        this.allDimensions = dataSourceConstraint.getAllDimensions();
+        this.allDimensionNames = dataSourceConstraint.getAllDimensionNames();
+        this.allColumnNames = dataSourceConstraint.getAllColumnNames();
+    }
+
     public Set<Dimension> getRequestDimensions() {
         return requestDimensions;
     }
@@ -84,5 +132,32 @@ public class DataSourceConstraint {
 
     public Map<Dimension, Set<ApiFilter>> getApiFilters() {
         return apiFilters;
+    }
+
+
+    /**
+     * Create a new <tt>DataSourceConstraint</tt> instance with a new subset of metric names.
+     * <p>
+     * The new set of metric names will be an intersection between old metric names and
+     * a user provided set of metric names
+     *
+     * @param metricNames  The set of metric names that are to be intersected with metric names in
+     * <tt>this DataSourceConstraint</tt>
+     *
+     * @return the new <tt>DataSourceConstraint</tt> instance with a new subset of metric names
+     */
+    public DataSourceConstraint withMetricIntersection(Set<String> metricNames) {
+        return new DataSourceConstraint(
+                requestDimensions,
+                filterDimensions,
+                metricDimensions,
+                metricNames.stream()
+                        .filter(this.metricNames::contains)
+                        .collect(Collectors.toSet()),
+                allDimensions,
+                allDimensionNames,
+                allColumnNames,
+                apiFilters
+        );
     }
 }