+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/config/table/BaseTableLoaderSpec.groovy
@@ -1,4 +1,4 @@
-// Copyright 2016 Yahoo Inc.
+// Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.data.config.table
 
@@ -8,31 +8,79 @@ import com.yahoo.bard.webservice.data.config.ResourceDictionaries
 import com.yahoo.bard.webservice.data.config.dimension.TestDimensions
 import com.yahoo.bard.webservice.data.config.names.ApiMetricName
 import com.yahoo.bard.webservice.data.config.names.FieldName
+import com.yahoo.bard.webservice.data.config.names.TableName
 import com.yahoo.bard.webservice.data.config.names.TestApiDimensionName
 import com.yahoo.bard.webservice.data.config.names.TestApiMetricName
 import com.yahoo.bard.webservice.data.config.names.TestDruidMetricName
 import com.yahoo.bard.webservice.data.dimension.Dimension
+import com.yahoo.bard.webservice.data.dimension.DimensionColumn
 import com.yahoo.bard.webservice.data.dimension.impl.KeyValueStoreDimension
+import com.yahoo.bard.webservice.data.time.DefaultTimeGrain
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService
-import com.yahoo.bard.webservice.metadata.TestDataSourceMetadataService
+import com.yahoo.bard.webservice.table.ConcretePhysicalTable
 import com.yahoo.bard.webservice.table.PhysicalTable
+import com.yahoo.bard.webservice.table.PhysicalTableSchema
 import com.yahoo.bard.webservice.table.TableGroup
+
+import org.joda.time.DateTimeZone
 
 import spock.lang.Specification
 
+import java.util.stream.Collectors
+
 /**
- * Testing basic table loader functionality
+ * Testing basic table loader functionality.
  */
 class BaseTableLoaderSpec extends Specification {
-    static class SimpleBaseTableLoader extends BaseTableLoader {
+    private static class SimpleBaseTableLoader extends BaseTableLoader {
 
-        public SimpleBaseTableLoader(DataSourceMetadataService metadataService) {
+        SimpleBaseTableLoader(DataSourceMetadataService metadataService) {
             super(metadataService)
         }
 
         @Override
-        public void loadTableDictionary(ResourceDictionaries dictionaries) {
+        void loadTableDictionary(ResourceDictionaries dictionaries) {
             // stub to allow testing of abstract class in unit tests
+        }
+    }
+
+    private static class SimpleDependencyPhysicalTableDefinition extends PhysicalTableDefinition {
+        String dependentTableName
+        PhysicalTable physicalTable
+
+        SimpleDependencyPhysicalTableDefinition(String tableName, String dependentTableName) {
+            super(TableName.of(tableName), DefaultTimeGrain.DAY.buildZonedTimeGrain(DateTimeZone.UTC), [] as Set, [] as Set)
+            this.dependentTableName = dependentTableName
+        }
+
+        SimpleDependencyPhysicalTableDefinition(String tableName, PhysicalTable physicalTable) {
+            super(TableName.of(tableName), DefaultTimeGrain.DAY.buildZonedTimeGrain(DateTimeZone.UTC), [] as Set, [] as Set)
+            this.physicalTable = physicalTable
+        }
+
+        @Override
+        Set<TableName> getDependentTableNames() {
+            if (Objects.isNull(dependentTableName)) {
+                return Collections.emptySet()
+            }
+            return Collections.singleton(TableName.of(dependentTableName));
+        }
+
+        @Override
+        PhysicalTable build(
+                ResourceDictionaries dictionaries,
+                DataSourceMetadataService metadataService
+        ) {
+            if (!Objects.isNull(physicalTable)) {
+                return physicalTable
+            }
+            return new ConcretePhysicalTable(
+                    TableName.of(getName().asName()),
+                    DefaultTimeGrain.DAY.buildZonedTimeGrain(DateTimeZone.UTC),
+                    [] as Set,
+                    [:],
+                    metadataService
+            )
         }
     }
 
@@ -43,28 +91,58 @@ class BaseTableLoaderSpec extends Specification {
     Set<PhysicalTableDefinition> physDefs
     Set<TestApiDimensionName> dimNames
     Collection<Dimension> dims
+    Set<TableName> tableNames
+    Set<PhysicalTableDefinition> allDependentDefinitions
+
+    PhysicalTableDefinition dependentDefinition1
+    PhysicalTableDefinition satisfiedDefinition2
+    PhysicalTableDefinition dependentDefinition3
+    PhysicalTableDefinition selfDependentDefinition4
+    PhysicalTableDefinition circularDependentDefinition5
+    PhysicalTableDefinition circularDependentDefinition6
+
+    PhysicalTable physicalTable
+    PhysicalTableSchema physicalTableSchema
 
     def setup() {
-        loader = new SimpleBaseTableLoader(new TestDataSourceMetadataService())
+        loader = new SimpleBaseTableLoader(Mock(DataSourceMetadataService))
         dicts = new ResourceDictionaries()
         apiNames = TestApiMetricName.getByLogicalTable(SHAPES)
         metricNames = TestDruidMetricName.getByLogicalTable(SHAPES)
-        physDefs = TestPhysicalTableDefinitionUtils.buildShapeTableDefinitions(new TestDimensions())
+        physDefs = TestPhysicalTableDefinitionUtils.buildShapeTableDefinitions(new TestDimensions(), metricNames)
         dimNames = TestApiDimensionName.getByLogicalTable(SHAPES)
+        tableNames = physDefs.collect {it.name} as Set
 
         dims = dimNames.collect {
             name -> new KeyValueStoreDimension(TestDimensions.buildStandardDimensionConfig(name))
         }
         dicts.getDimensionDictionary().addAll(dims)
+
+        physicalTableSchema = Mock(PhysicalTableSchema)
+        physicalTableSchema.getColumns(DimensionColumn.class) >> []
+
+        physicalTable = Mock(PhysicalTable)
+        physicalTable.getTableName() >> TableName.of('definition2')
+        physicalTable.getSchema() >> physicalTableSchema
+
+        dependentDefinition1 = new SimpleDependencyPhysicalTableDefinition('definition1', 'definition2')
+        satisfiedDefinition2 = new SimpleDependencyPhysicalTableDefinition('definition2', physicalTable)
+        dependentDefinition3 = new SimpleDependencyPhysicalTableDefinition('definition3', 'definition1')
+        selfDependentDefinition4 = new SimpleDependencyPhysicalTableDefinition('definition4', 'definition4')
+        circularDependentDefinition5 = new SimpleDependencyPhysicalTableDefinition('definition5', 'definition6')
+        circularDependentDefinition6 = new SimpleDependencyPhysicalTableDefinition('definition6', 'definition5')
+
+        allDependentDefinitions = [dependentDefinition1, satisfiedDefinition2, dependentDefinition3, selfDependentDefinition4, circularDependentDefinition5, circularDependentDefinition6]
+
     }
 
     def "table group has correct contents after being build"() {
         when:
         TableGroup group = loader.buildDimensionSpanningTableGroup(
-                apiNames,
-                metricNames,
+                tableNames,
                 physDefs,
-                dicts
+                dicts,
+                apiNames
         )
 
         then:
@@ -72,31 +150,124 @@ class BaseTableLoaderSpec extends Specification {
         group.apiMetricNames == apiNames
     }
 
-    def "loading distinct physical tables results in correct tables in dictionary"() {
-        setup:
-        List<PhysicalTableDefinition> allDefs = physDefs as List
 
+    def "loading distinct physical tables without dependency results in correct tables in dictionary and table group"() {
         when:
-        PhysicalTable t1 = loader.loadPhysicalTable(allDefs[0], metricNames, dicts)
-        PhysicalTable t2 = loader.loadPhysicalTable(allDefs[1], metricNames, dicts)
+        TableGroup group = loader.buildDimensionSpanningTableGroup(
+                tableNames,
+                physDefs,
+                dicts,
+                apiNames
+        )
 
         then:
-        dicts.physicalDictionary.size() == 2
-        dicts.physicalDictionary.containsValue(t1)
-        dicts.physicalDictionary.containsValue(t2)
+        dicts.physicalDictionary.size() == 6
+        group.physicalTables.collect {it.getTableName()} as Set == tableNames
+    }
+
+    def "loading physical tables with dependency loads all satisfied dependency physical tables"() {
+        given:
+        Set<TableName> currentTableNames = [dependentDefinition1.name, satisfiedDefinition2.name, dependentDefinition3.name]
+
+        when:
+        TableGroup group = loader.buildDimensionSpanningTableGroup(
+                currentTableNames,
+                allDependentDefinitions,
+                dicts,
+                apiNames
+        )
+
+        then:
+        dicts.physicalDictionary.size() == 3
+        group.physicalTables.collect {it.getTableName()} as Set == currentTableNames
+    }
+
+    def "loading a physical table with dependency outside of the current table group will be loaded successfully"() {
+        given:
+        Set<TableName> currentTableNames = [satisfiedDefinition2.name, dependentDefinition3.name]
+
+        when:
+        TableGroup group = loader.buildDimensionSpanningTableGroup(
+                currentTableNames,
+                allDependentDefinitions,
+                dicts,
+                apiNames
+        )
+
+        then:
+        dicts.physicalDictionary.size() == 3
+        group.physicalTables.collect { it.getTableName() } as Set == currentTableNames
+    }
+
+    def "unsatisfied dependency physical table definition loading will throw an exception"() {
+        given:
+        Set<TableName> currentTableNames = [satisfiedDefinition2.name, dependentDefinition3.name]
+
+        when:
+        loader.buildDimensionSpanningTableGroup(
+                currentTableNames,
+                [satisfiedDefinition2, dependentDefinition3] as Set,
+                dicts,
+                apiNames
+        )
+
+        then:
+        RuntimeException e = thrown()
+        e.message == 'Unable to resolve physical table dependency for physical table: definition1, might be missing or circular dependency'
+    }
+
+    def "circular dependency physical table definition loading will throw an exception"() {
+        given:
+        Set<TableName> currentTableNames = [dependentDefinition1.name, satisfiedDefinition2.name, dependentDefinition3.name, circularDependentDefinition5.name, circularDependentDefinition6.name]
+
+        when:
+        loader.buildDimensionSpanningTableGroup(
+                currentTableNames,
+                allDependentDefinitions,
+                dicts,
+                apiNames
+        )
+
+        then:
+        RuntimeException e = thrown()
+        e.message == 'Unable to resolve physical table dependency for physical table: definition5, might be missing or circular dependency'
+    }
+
+    def "self dependency physical table definition loading will throw an exception"() {
+        given:
+        Set<TableName> currentTableNames = [dependentDefinition1.name, satisfiedDefinition2.name, dependentDefinition3.name, selfDependentDefinition4.name]
+
+        when:
+        loader.buildDimensionSpanningTableGroup(
+                currentTableNames,
+                allDependentDefinitions,
+                dicts,
+                apiNames
+        )
+
+        then:
+        RuntimeException e = thrown()
+        e.message == 'Unable to resolve physical table dependency for physical table: definition4, might be missing or circular dependency'
     }
 
     def "load duplicate physical tables results in sharing definitions"() {
-        setup:
-        List<PhysicalTableDefinition> allDefs = physDefs as List
-
         when:
-        PhysicalTable t1 = loader.loadPhysicalTable(allDefs[0], metricNames, dicts)
-        PhysicalTable t2 = loader.loadPhysicalTable(allDefs[0], metricNames, dicts)
+        TableGroup group1 = loader.buildDimensionSpanningTableGroup(
+                tableNames,
+                physDefs,
+                dicts,
+                apiNames
+        )
+        TableGroup group2 = loader.buildDimensionSpanningTableGroup(
+                tableNames,
+                physDefs,
+                dicts,
+                apiNames
+        )
 
         then:
-        dicts.physicalDictionary.size() == 1
-        dicts.physicalDictionary.containsValue(t1)
-        t1.is(t2)
+        dicts.physicalDictionary.size() == 6
+        group1.physicalTables.collect {it.getTableName()} as Set == tableNames
+        group2.physicalTables.collect {it.getTableName()} as Set == tableNames
     }
 }